# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "鸵小妥API",
        "contact": {
            "email": "wq0426@163.com"
        },
        "version": "1.0.2"
    },
    "host": "127.0.0.1:8289",
    "paths": {
        "/account/code": {
            "post": {
                "description": "发送验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户模块"
                ],
                "summary": "发送验证码",
                "parameters": [
                    {
                        "description": "备注 (type 1:注册 2:登录 3:找回密码 4:修改手机号 5:重新绑定 6:交友密码)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.CryptoRequest"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/v1.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "boolean"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/news/list": {
            "get": {
                "description": "获取新闻资讯列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "新闻模块"
                ],
                "summary": "获取新闻资讯列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.NewsType"
                            }
                        }
                    }
                }
            }
        },
...
        "/refund/cancel": {
            "post": {
                "description": "撤销处理中的退款申请，恢复原订单状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "退款"
                ],
                "summary": "撤销退款申请",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "退款撤销信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CancelRefundRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "403": {
                        "description": "退款单不存在或无权限",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "422": {
                        "description": "当前退款单状态不允许撤销",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/refund/create": {
            "post": {
                "description": "创建退款申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "退款"
                ],
                "summary": "创建退款单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "创建退款请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RefundItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/refund/delete/{refund_id}": {
            "delete": {
                "description": "删除用户的退款记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "退款"
                ],
                "summary": "删除退款记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "退款单ID",
                        "name": "refund_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "403": {
                        "description": "退款单不存在或无权限",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/refund/detail/{refund_id}": {
            "get": {
                "description": "获取退款单详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "退款"
                ],
                "summary": "获取退款单详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "退款单ID",
                        "name": "refund_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RefundDetailResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "404": {
                        "description": "退款单不存在",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
        "/refund/list": {
            "get": {
                "description": "获取用户退款单列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "退款"
                ],
                "summary": "获取退款单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "退款状态：0-待处理，1-退款中，2-已退款，3-已拒绝；不传则查询所有状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码，默认1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数，默认10",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RefundListResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CancelRefundRequest": {
            "type": "object",
            "required": [
                "refund_id"
            ],
            "properties": {
                "refund_id": {
                    "description": "退款单ID",
                    "type": "integer"
                }
            }
        },
...
        "model.NewsResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "date": {
                    "description": "日期",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "img": {
                    "description": "图片链接",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "url": {
                    "description": "跳转链接",
                    "type": "string"
                }
            }
        },
        "model.NewsType": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "新闻列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NewsResponse"
                    }
                },
                "title": {
                    "description": "类型标题",
                    "type": "string"
                },
                "type": {
                    "description": "类型",
                    "type": "integer"
                }
            }
        },
...
        "model.OrderDetails": {
            "type": "object",
            "properties": {
                "order_created": {
                    "description": "订单创建时间",
                    "type": "string"
                },
                "order_id": {
                    "description": "订单ID",
                    "type": "integer"
                },
                "order_no": {
                    "description": "订单号",
                    "type": "string"
                },
                "order_status": {
                    "description": "订单状态",
                    "type": "integer"
                }
            }
        },
...
        "model.RefundDetailResponse": {
            "type": "object",
            "properties": {
                "order_detail": {
                    "description": "订单详情",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.OrderDetails"
                        }
                    ]
                },
                "refund": {
                    "description": "退款单信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.RefundOrder"
                        }
                    ]
                }
            }
        },
        "model.RefundItemRequest": {
            "type": "object",
            "required": [
                "order_item_id",
                "refund_reason",
                "refund_type"
            ],
            "properties": {
                "images": {
                    "description": "图片凭证",
                    "type": "string"
                },
                "order_item_id": {
                    "description": "订单项ID",
                    "type": "integer"
                },
                "refund_reason": {
                    "description": "退款原因",
                    "type": "string"
                },
                "refund_type": {
                    "description": "退款类型(1:仅退款;2:退货退款)",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                }
            }
        },
        "model.RefundListResponse": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "退款列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RefundOrderItem"
                    }
                },
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "size": {
                    "description": "每页条数",
                    "type": "integer"
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "model.RefundOrder": {
            "type": "object",
            "properties": {
                "apply_time": {
                    "description": "申请时间",
                    "type": "string"
                },
                "completion_time": {
                    "description": "完成时间",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "header_img": {
                    "description": "商品图片",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "description": "图片凭证",
                    "type": "string"
                },
                "order_id": {
                    "description": "订单ID",
                    "type": "integer"
                },
                "order_item_id": {
                    "description": "订单项ID",
                    "type": "integer"
                },
                "order_no": {
                    "description": "订单号",
                    "type": "string"
                },
                "origin_status": {
                    "description": "原始状态",
                    "type": "integer"
                },
                "price": {
                    "description": "商品单价",
                    "type": "number"
                },
                "process_time": {
                    "description": "处理时间",
                    "type": "string"
                },
                "product_id": {
                    "description": "商品ID",
                    "type": "integer"
                },
                "product_name": {
                    "description": "商品名称",
                    "type": "string"
                },
                "quantity": {
                    "description": "退款数量",
                    "type": "integer"
                },
                "refund_amount": {
                    "description": "退款金额",
                    "type": "number"
                },
                "refund_no": {
                    "description": "退款单号",
                    "type": "string"
                },
                "refund_reason": {
                    "description": "退款原因",
                    "type": "string"
                },
                "refund_type": {
                    "description": "退款类型",
                    "type": "integer"
                },
                "reject_reason": {
                    "description": "拒绝原因",
                    "type": "string"
                },
                "status": {
                    "description": "退款状态",
                    "type": "integer"
                },
                "store_icon": {
                    "description": "店铺图标",
                    "type": "string"
                },
                "store_name": {
                    "description": "店铺名称",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string"
                }
            }
        },
        "model.RefundOrderItem": {
            "type": "object",
            "properties": {
                "apply_time": {
                    "description": "申请时间",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "header_img": {
                    "description": "商品图片",
                    "type": "string"
                },
                "id": {
                    "description": "退款单ID",
                    "type": "integer"
                },
                "order_id": {
                    "description": "订单ID",
                    "type": "integer"
                },
                "order_no": {
                    "description": "订单号",
                    "type": "string"
                },
                "price": {
                    "description": "商品单价",
                    "type": "number"
                },
                "product_name": {
                    "description": "商品名称",
                    "type": "string"
                },
                "quantity": {
                    "description": "退款数量",
                    "type": "integer"
                },
                "refund_amount": {
                    "description": "退款金额",
                    "type": "number"
                },
                "refund_no": {
                    "description": "退款单号",
                    "type": "string"
                },
                "refund_type": {
                    "description": "退款类型",
                    "type": "integer"
                },
                "status": {
                    "description": "退款状态",
                    "type": "integer"
                },
                "status_text": {
                    "description": "退款状态文本",
                    "type": "string"
                },
                "store_icon": {
                    "description": "店铺图标",
                    "type": "string"
                },
                "store_name": {
                    "description": "店铺名称",
                    "type": "string"
                }
            }
        },
...
        "v1.CryptoRequest": {
            "description": "CryptoRequest 参数描述",
            "type": "object",
            "required": [
                "phone"
            ],
            "properties": {
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "v1.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}